backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.9
pydantic==1.8.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
alembic==1.7.1
scikit-learn==1.2.2
pandas==2.0.1
numpy==1.24.3
pydotplus==2.0.2
graphviz==0.20.1
jsonify==0.5
openpyxl==3.1.2
matplotlib==3.7.1
PyJWT==2.8.0

# Development and Testing Dependencies
pytest==7.4.0
httpx==0.24.1

# New dependencies
fastapi-cache2[redis]==0.2.1
redis>=4.2.0rc1,<5.0.0